a bit of organization on the structure tests
