+++ system/ee/ExpressionEngine/Tests/ExpressionEngine/Addons/Structure/Mod/StructureRemoveLastSegmentTest.php
