+++ system/ee/ExpressionEngine/Tests/ExpressionEngine/Addons/Structure/Mod/Variables/StructureChildIdsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once __DIR__ . '/StructureTestBase.php';
+require_once __DIR__ . '/../StructureTestBase.php';
 
 class StructureChildIdsTest extends StructureTestBase
 {
