+++ system/ee/ExpressionEngine/Tests/ExpressionEngine/Addons/Structure/Mod/tags/StructureFirstChildRedirectTest.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once __DIR__ . '/StructureTestBase.php';
+require_once __DIR__ . '/../StructureTestBase.php';
 
 class StructureFirstChildRedirectTest extends StructureTestBase
 {
